20210731 code dump.

+++ Common/PokemonSwSh/PokemonSwShDateSpam.cpp
@@ -174,7 +174,7 @@ int register_message_converters_pokemon_date_spam(){
             if (body.size() != sizeof(pabb_rollback_hours_from_home)){ ss << "(invalid size)" << std::endl; return ss.str(); }
             const auto* params = (const pabb_rollback_hours_from_home*)body.c_str();
             ss << "seqnum = " << (uint64_t)params->seqnum;
-            ss << ", hours = " << params->hours;
+            ss << ", hours = " << (int)params->hours;
             ss << ", settings_to_home_delay = " << params->settings_to_home_delay;
             return ss.str();
         }
